fix claim surface area overflowing int on huge regions
getSurfaceArea() multiplied two ints, so a region near the +/-32,000,000
coordinate limit wrapped around to a small or negative number. that let
createClaimAt/resizeClaim's claim-block cost checks pass for free since
"blocks < surfaceArea" is never true against a negative area, letting
someone claim or extend a claim to a huge size without the blocks to
pay for it. do the math in long instead.

+++ common/src/main/java/net/william278/huskclaims/claim/ClaimWorld.java
@@ -234,7 +234,7 @@ public class ClaimWorld {
      * @since 1.0
      */
     public long getSurfaceClaimedBy(@NotNull User owner) {
-        return getClaimsByUser(owner.getUuid()).stream().mapToInt(c -> c.getRegion().getSurfaceArea()).sum();
+        return getClaimsByUser(owner.getUuid()).stream().mapToLong(c -> c.getRegion().getSurfaceArea()).sum();
     }
 
     /**
